+++ frontend/src/pages/Account.jsx
@@ -1,120 +1,63 @@
-import React, { useEffect, useState } from "react";
-import SelectInput from "../components/SelectInput/SelectInput";
+import { useEffect, useState } from "react";
+import { Link, useParams } from "react-router-dom";
+import Profile from "../components/Account/Profile";
 import AuthService from "../utils/AuthService";
-import FacultyService from "../utils/FacultyService";
-import styles from "./Forms.module.css";
+import styles from "./Account.module.css";
 
 function Account() {
-  const [name, setName] = useState();
-  const [email, setEmail] = useState();
-  const [faculty, setFaculty] = useState();
-  const [department, setDepartment] = useState();
-  let faculties;
-  let departments;
+  const { section } = useParams();
+  const [selectedSection, setSelectedSection] = useState("profile");
 
-  const getFaculties = async () => {
-    const response = await FacultyService.getFaculties();
-    return response;
-  };
-
-  const getDepartments = async () => {
-    const response = await FacultyService.getDepartments();
-    return response;
-  };
-
-  const handleUpdateUser = async () => {
-    /*     if (!name || !email || !faculty || !department) {
-      setErrorMessage("Please provide all the required fields.");
-      return;
+  const displaySelection = (selection) => {
+    switch (selection) {
+      case "Profile":
+        return <Profile />;
+      case "Settings":
+        return null;
+      default:
+        return <Profile />;
     }
-    const response = await AuthService.updateUser(
-      user.id,
-      name,
-      email,
-      faculty.id,
-      department.id
-    );
-    if (response.status && response.status !== 200) {
-      setMessage(undefined);
-      setErrorMessage(response.data.message);
-    } else {
-      setErrorMessage(undefined);
-      setMessage(response.message);
-    } */
   };
 
-  const retrieveUser = async (faculties, departments) => {
-    const response = await AuthService.getUser();
-    setEmail(response.email);
-    setName(response.name);
-    const userFaculty = faculties.find(
-      (faculty) => faculty.id === response.faculty_id
-    );
-    setFaculty(userFaculty);
-    const userDepartment = departments.find(
-      (department) => department.id === response.department_id
-    );
-    setDepartment(userDepartment);
-  };
-
-  const getData = async () => {
-    faculties = await getFaculties();
-    departments = await getDepartments();
-    await retrieveUser(faculties, departments);
+  const convertToTitleCase = (string) => {
+    return string.charAt(0).toUpperCase() + string.substr(1).toLowerCase();
   };
 
   useEffect(() => {
-    getData();
-  }, []);
+    setSelectedSection(convertToTitleCase(section));
+  }, [section]);
 
   return (
-    <div>
-      <div className={styles.formContainer}>
-        <div className={styles.formTitle}>User Info</div>
-        <div className={styles.formSection}>
-          <p>Name</p>
-          <input
-            type="text"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            className={styles.formInput}
-          />
-        </div>
-        <div className={styles.formSection}>
-          <p>Email</p>
-          <input
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            className={styles.formInput}
-          />
-        </div>
-        <div className={styles.formSection}>
-          <p>Faculty</p>
-          <SelectInput
-            value={faculty}
-            items={faculties}
-            selectItem={setFaculty}
-            hint="Loading..."
-          />
-        </div>
-        <div className={styles.formSection}>
-          <p>Department</p>
-          <SelectInput
-            value={department}
-            items={departments}
-            selectItem={setDepartment}
-            hint="Choose a department"
-          />
-        </div>
-        <button
-          type="submit"
-          className={styles.formSubmitButton}
-          onClick={() => handleUpdateUser()}
+    <div className={styles.accountContainer}>
+      <div className={styles.sectionButtons}>
+        <Link
+          to={"/account/profile"}
+          className={
+            selectedSection == "Profile"
+              ? styles.sectionButtonActive
+              : styles.sectionButton
+          }
         >
-          Save Changes
+          Profile
+        </Link>
+        <Link
+          to={"/account/settings"}
+          className={
+            selectedSection == "Settings"
+              ? styles.sectionButtonActive
+              : styles.sectionButton
+          }
+        >
+          Settings
+        </Link>
+        <button
+          className={styles.logOutButton}
+          onClick={() => AuthService.logout()}
+        >
+          LOG OUT
         </button>
       </div>
+      {displaySelection(selectedSection)}
     </div>
   );
 }
